--- infrastructure/src/main/java/org/hesperides/infrastructure/jpa/modules/commands/JpaModuleCommandsRepository.java
@@ -1,75 +1,0 @@
-package org.hesperides.infrastructure.jpa.modules.commands;
-
-import org.axonframework.eventsourcing.EventSourcingHandler;
-import org.hesperides.domain.modules.ModuleCreatedEvent;
-import org.hesperides.domain.modules.ModuleDeletedEvent;
-import org.hesperides.domain.modules.ModuleUpdatedEvent;
-import org.hesperides.domain.modules.commands.ModuleCommandsRepository;
-import org.hesperides.domain.modules.entities.Module;
-import org.hesperides.domain.templatecontainer.entities.TemplateContainer;
-import org.hesperides.infrastructure.jpa.modules.JpaModuleRepository;
-import org.hesperides.infrastructure.jpa.modules.ModuleEntity;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Component;
-
-import static org.hesperides.domain.Profiles.JPA;
-
-@Profile(JPA)
-@Component
-public class JpaModuleCommandsRepository implements ModuleCommandsRepository {
-
-    private final JpaModuleRepository jpaModuleRepository;
-
-    @Autowired
-    public JpaModuleCommandsRepository(JpaModuleRepository jpaModuleRepository) {
-        this.jpaModuleRepository = jpaModuleRepository;
-    }
-
-    @EventSourcingHandler
-    @Override
-    public void on(ModuleCreatedEvent event) {
-        Module module = event.getModule();
-        ModuleEntity.ModuleEntityId id = new ModuleEntity.ModuleEntityId(
-                module.getKey().getName(),
-                module.getKey().getVersion(),
-                module.getKey().getVersionType()
-        );
-        jpaModuleRepository.save(
-                new ModuleEntity(
-                        id,
-                        module.getVersionId()
-                )
-        );
-    }
-
-    @EventSourcingHandler
-    @Override
-    public void on(ModuleUpdatedEvent event) {
-        Module module = event.getModule();
-        ModuleEntity.ModuleEntityId id = new ModuleEntity.ModuleEntityId(
-                module.getKey().getName(),
-                module.getKey().getVersion(),
-                module.getKey().getVersionType()
-        );
-        jpaModuleRepository.save(
-                new ModuleEntity(
-                        id,
-                        module.getVersionId()
-                )
-        );
-    }
-
-    @EventSourcingHandler
-    @Override
-    public void on(ModuleDeletedEvent event) {
-        TemplateContainer.Key key = event.getModuleKey();
-        ModuleEntity.ModuleEntityId id = new ModuleEntity.ModuleEntityId(
-                key.getName(),
-                key.getVersion(),
-                key.getVersionType()
-        );
-        jpaModuleRepository.delete(id);
-    }
-
-}
