+++ infrastructure/src/main/java/org/hesperides/infrastructure/mongo/modules/MongoModuleProjectionRepository.java
@@ -1,12 +1,13 @@
-package org.hesperides.infrastructure.mongo.modules.queries;
+package org.hesperides.infrastructure.mongo.modules;
 
+import org.axonframework.eventsourcing.EventSourcingHandler;
 import org.axonframework.queryhandling.QueryHandler;
 import org.hesperides.domain.modules.*;
-import org.hesperides.domain.modules.queries.ModuleQueriesRepository;
+import org.hesperides.domain.modules.entities.Module;
 import org.hesperides.domain.modules.queries.ModuleView;
 import org.hesperides.domain.templatecontainer.entities.TemplateContainer;
-import org.hesperides.infrastructure.mongo.modules.ModuleDocument;
-import org.hesperides.infrastructure.mongo.modules.MongoModuleRepository;
+import org.hesperides.infrastructure.mongo.technos.MongoTechnoProjectionRepository;
+import org.hesperides.infrastructure.mongo.technos.TechnoDocument;
 import org.hesperides.infrastructure.mongo.templatecontainer.KeyDocument;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
@@ -23,16 +24,48 @@ import static org.hesperides.domain.Profiles.*;
 
 @Profile({MONGO, EMBEDDED_MONGO, FAKE_MONGO})
 @Repository
-public class MongoModuleQueriesRepository implements ModuleQueriesRepository {
+public class MongoModuleProjectionRepository implements ModuleProjectionRepository {
 
     private final MongoModuleRepository moduleRepository;
+    private final MongoTechnoProjectionRepository technoProjectionRepository;
     private final MongoTemplate mongoTemplate;
 
     @Autowired
-    public MongoModuleQueriesRepository(MongoModuleRepository moduleRepository, MongoTemplate mongoTemplate) {
+    public MongoModuleProjectionRepository(MongoModuleRepository moduleRepository,
+                                           MongoTechnoProjectionRepository technoProjectionRepository,
+                                           MongoTemplate mongoTemplate) {
         this.moduleRepository = moduleRepository;
+        this.technoProjectionRepository = technoProjectionRepository;
         this.mongoTemplate = mongoTemplate;
     }
+
+    /*** EVENT HANDLERS ***/
+
+    @EventSourcingHandler
+    @Override
+    public void on(ModuleCreatedEvent event) {
+        saveFromDomainInstance(event.getModule());
+    }
+
+    @EventSourcingHandler
+    @Override
+    public void on(ModuleUpdatedEvent event) {
+        saveFromDomainInstance(event.getModule());
+    }
+
+    private void saveFromDomainInstance(Module module) {
+        List<TechnoDocument> technoDocuments = technoProjectionRepository.getTechnoDocumentsFromDomainInstances(module.getTechnos());
+        ModuleDocument moduleDocument = ModuleDocument.fromDomainInstance(module, technoDocuments);
+        moduleRepository.save(moduleDocument);
+    }
+
+    @EventSourcingHandler
+    @Override
+    public void on(ModuleDeletedEvent event) {
+        moduleRepository.deleteByKey(KeyDocument.fromDomainInstance(event.getModuleKey()));
+    }
+
+    /*** QUERY HANDLERS ***/
 
     @QueryHandler
     @Override
